+++ release-notes.md
@@ -7,6 +7,7 @@
 
 * Submit timestamps to multiple calendars in parallel.
 * git-extract subcommand now works with relative paths.
+* Improve error message when git-extract used on a non-rehash-trees commit.
 
 
 ## v0.2.3